PERF: Use backend to load `Note` (#14948)
* log: remove 'note.load'

Added in 19e92c14de6bc7f052dcf389e574e76f9a614968

Doesn't add much value, causes spam in CardBrowser

* refactor: Kotlin: note.tags -> MutableList

Removes a cast. Backend uses an **immutable** list
 but this is not an ArrayList, and we still want this
 value to be mutable due to other methods on the API surface

* refactor: Kotlin: note.fields -> MutableList

Removes a cast. Backend uses an **immutable** list
 but this is not an ArrayList, and we still want this
 value to be mutable due to other methods on the API surface

* refactor: Kotlin: note.mod -> Int

This is what the backend uses:

anki.notes.Note.getMtimeSecs is an uint32

* refactor: note: remove 'scm'

unused

* refactor: note: remove 'flags'

unused

* refactor: note: remove 'data'

unused

* tests: fix - use RunWith(AndroidJUnit4)

This may have caused issue 14796, as the inheritor was inheriting from RobolectricTest()

After fixing `Note` to use the backend, I got one
 non-reproducible error:

 `java.lang.UnsatisfiedLinkError: 'byte[][] net.ankiweb.rsdroid.NativeMethods.openBackend(byte[])'`

I strongly suspect that it was this class which caused
 the issue.

If this is incorrect, it's a small tiem penalty to unit tests

* note: load from Backend

This reduced the time to load 1 million notes from ~1m30 -> ~30s

https://github.com/ankitects/anki/blob/db93939ded947d74bb25bd8552a2b2356a096509/pylib/anki/notes.py#L54-L67

+++ AnkiDroid/src/main/java/com/ichi2/anki/CardBrowser.kt
@@ -2019,7 +2019,7 @@ open class CardBrowser :
                 Column.CARD -> if (inCardMode) card.template().optString("name") else "${card.note().numberOfCards()}"
                 Column.DUE -> card.dueString
                 Column.EASE -> if (inCardMode) getEaseForCards() else getAvgEaseForNotes()
-                Column.CHANGED -> LanguageUtil.getShortDateFormatFromS(if (inCardMode) card.mod else card.note().mod)
+                Column.CHANGED -> LanguageUtil.getShortDateFormatFromS(if (inCardMode) card.mod else card.note().mod.toLong())
                 Column.CREATED -> LanguageUtil.getShortDateFormatFromMs(card.note().id)
                 Column.EDITED -> LanguageUtil.getShortDateFormatFromS(card.note().mod)
                 Column.INTERVAL -> if (inCardMode) queryIntervalForCards() else queryAvgIntervalForNotes()
